Establish initial ndb models and supporting infrastructure (WIP)
Set up some initial tests (driven by py.test) to confirm workings
of ndb models and grammar for queries. This means we need to
make sure:

* that the environ has the necessary python modules on the sys.path
* that the os ENV has the right things

This is done with code in lib/__init__.py (cribbed from the
djappengine sample) tests/__init__.py, and test/fixtures.py (based
on appengine testing docs).

The urls and views are not doing anything yet, just there as
boilerplate, I determined it was best to make sure that the
datastore could be made to work before moving on.

manage.py, settings.py and shell.py are adapted from djappengine.
Before this stuff is done those files will need to be made more
portable to other environments.