Support creating a blog
No support yet for entries, just the blog objects themself.

Quite a lot of changes in this commit as this work was done
operating off the test server rather than using tests, largely
because I was working on UI issues in an exploratory way.

The Blog model class now has a get_absoluate_url so we can
redirect with a blog object.

The editor and creator routes need to go before the blog and
entry routes in urls.py.

To create a blog we must be able to log in, so now we have that.