Now possible to create a new entry in an existing blog.
The tricks here were figuring out how to:

* correctly validate forms
* how to pre-populate a form
* understanding ndb.Key

The style for the form(s) is very bad, but that can be easily
improved with a bit of work. I want to make sure I get the
remaining desired functionality in first.

Again: most of this work was driven without tests because of
exploring UI and the tools. If I were more familiar with the tools
then more testing would be the norm.

Next up: editing existing entries.