Delete a single entry from its display page.
A small form is placed on the page with a button and label for
submit. The button is moved off screen so that the label (an icon)
can be the main actor. JavaScript is attached to the form to confirm
the delete.

All this trouble is being done to insure that the delete is done
with an unsafe method (in this case POST) and will also still work
if JavaScript is turned off.