Treat entry content as Markdown
Render the markdown via a property on the Model which
calls to render:render where the Markdown library does its
stuff.

Markdown safe mode is set.

The markdown library is pulled in via a new Makefile target
that calls a short addlib.py script to _symlink_ the desired
module in.